requirements.txt
moviepy==1.0.3
Pillow>=10.0.0
numpy>=1.24.0
streamlit>=1.30.0
